Add regVersion option to specify 32 or 64 bit registry
Add regVersion option in order to specify 32 or 64 bits is the key. Corresponds to REG's /REG:32 and /REG:64 options. Add validation that passed value is 32 or 64. If value is not passed, regVersion is not passed to REG command line tool.
Should be used in the following way:

var Winreg = require('winreg')
,   regKey = new Winreg({
      hive: Winreg.HKLM,                                          // HKEY_CURRENT_USER
      key: '\\Software\\Microsoft\\Windows\\CurrentVersion\\Run', // key containing autostart programs
      regVersion: '32',  // if you specify 64 you'll see different result
    })